+++ src/RcppExports.cpp
@@ -29,6 +29,24 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
+// cpp_register_occu_cover_lognormal_agg_coupling
+void cpp_register_occu_cover_lognormal_agg_coupling();
+RcppExport SEXP _tulpaObs_cpp_register_occu_cover_lognormal_agg_coupling() {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    cpp_register_occu_cover_lognormal_agg_coupling();
+    return R_NilValue;
+END_RCPP
+}
+// cpp_register_occu_cover_beta_agg_coupling
+void cpp_register_occu_cover_beta_agg_coupling();
+RcppExport SEXP _tulpaObs_cpp_register_occu_cover_beta_agg_coupling() {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    cpp_register_occu_cover_beta_agg_coupling();
+    return R_NilValue;
+END_RCPP
+}
 // cpp_eval_occu_cover_lognormal_cell
 Rcpp::List cpp_eval_occu_cover_lognormal_cell(double eta_psi, Rcpp::NumericVector eta_p, Rcpp::NumericVector eta_pos, Rcpp::IntegerVector y_det, Rcpp::NumericVector y_pos, double sigma_pos, std::string curvature);
 RcppExport SEXP _tulpaObs_cpp_eval_occu_cover_lognormal_cell(SEXP eta_psiSEXP, SEXP eta_pSEXP, SEXP eta_posSEXP, SEXP y_detSEXP, SEXP y_posSEXP, SEXP sigma_posSEXP, SEXP curvatureSEXP) {
@@ -60,6 +78,40 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type phi_pos(phi_posSEXP);
     Rcpp::traits::input_parameter< std::string >::type curvature(curvatureSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_eval_occu_cover_beta_cell(eta_psi, eta_p, eta_pos, y_det, y_pos, phi_pos, curvature));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_eval_occu_cover_lognormal_agg_cell
+Rcpp::List cpp_eval_occu_cover_lognormal_agg_cell(double eta_psi, Rcpp::NumericVector eta_p, Rcpp::NumericVector eta_pos, Rcpp::IntegerVector y_det, Rcpp::NumericVector y_pos, double sigma_pos, std::string curvature);
+RcppExport SEXP _tulpaObs_cpp_eval_occu_cover_lognormal_agg_cell(SEXP eta_psiSEXP, SEXP eta_pSEXP, SEXP eta_posSEXP, SEXP y_detSEXP, SEXP y_posSEXP, SEXP sigma_posSEXP, SEXP curvatureSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type eta_psi(eta_psiSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type eta_p(eta_pSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type eta_pos(eta_posSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type y_det(y_detSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type y_pos(y_posSEXP);
+    Rcpp::traits::input_parameter< double >::type sigma_pos(sigma_posSEXP);
+    Rcpp::traits::input_parameter< std::string >::type curvature(curvatureSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_eval_occu_cover_lognormal_agg_cell(eta_psi, eta_p, eta_pos, y_det, y_pos, sigma_pos, curvature));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_eval_occu_cover_beta_agg_cell
+Rcpp::List cpp_eval_occu_cover_beta_agg_cell(double eta_psi, Rcpp::NumericVector eta_p, Rcpp::NumericVector eta_pos, Rcpp::IntegerVector y_det, Rcpp::NumericVector y_pos, double phi_pos, std::string curvature);
+RcppExport SEXP _tulpaObs_cpp_eval_occu_cover_beta_agg_cell(SEXP eta_psiSEXP, SEXP eta_pSEXP, SEXP eta_posSEXP, SEXP y_detSEXP, SEXP y_posSEXP, SEXP phi_posSEXP, SEXP curvatureSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type eta_psi(eta_psiSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type eta_p(eta_pSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type eta_pos(eta_posSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type y_det(y_detSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type y_pos(y_posSEXP);
+    Rcpp::traits::input_parameter< double >::type phi_pos(phi_posSEXP);
+    Rcpp::traits::input_parameter< std::string >::type curvature(curvatureSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_eval_occu_cover_beta_agg_cell(eta_psi, eta_p, eta_pos, y_det, y_pos, phi_pos, curvature));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -559,8 +611,12 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_tulpaObs_cpp_register_occu_cover_lognormal_coupling", (DL_FUNC) &_tulpaObs_cpp_register_occu_cover_lognormal_coupling, 0},
     {"_tulpaObs_cpp_register_occu_cover_beta_coupling", (DL_FUNC) &_tulpaObs_cpp_register_occu_cover_beta_coupling, 0},
+    {"_tulpaObs_cpp_register_occu_cover_lognormal_agg_coupling", (DL_FUNC) &_tulpaObs_cpp_register_occu_cover_lognormal_agg_coupling, 0},
+    {"_tulpaObs_cpp_register_occu_cover_beta_agg_coupling", (DL_FUNC) &_tulpaObs_cpp_register_occu_cover_beta_agg_coupling, 0},
     {"_tulpaObs_cpp_eval_occu_cover_lognormal_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_cover_lognormal_cell, 7},
     {"_tulpaObs_cpp_eval_occu_cover_beta_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_cover_beta_cell, 7},
+    {"_tulpaObs_cpp_eval_occu_cover_lognormal_agg_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_cover_lognormal_agg_cell, 7},
+    {"_tulpaObs_cpp_eval_occu_cover_beta_agg_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_cover_beta_agg_cell, 7},
     {"_tulpaObs_cpp_register_occu_multiscale_cover_coupling", (DL_FUNC) &_tulpaObs_cpp_register_occu_multiscale_cover_coupling, 3},
     {"_tulpaObs_cpp_eval_occu_multiscale_cover_lognormal_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_multiscale_cover_lognormal_cell, 9},
     {"_tulpaObs_cpp_eval_occu_multiscale_cover_beta_cell", (DL_FUNC) &_tulpaObs_cpp_eval_occu_multiscale_cover_beta_cell, 9},